conditions

+++ src/package_name/AmplitudeReductionEffects.py
@@ -94,18 +94,20 @@ def add_amplitude_reduction_effects(reference_dir: str, volume_threshold: float)
 
     print("\033[92mVolume Reduced successfully!\033[0m")
 
-    # Create a text file to store the output audio files
-    os.chdir(reference_dir)
-    os.chdir("../")
+    if len(output_files) > 0:
+        # Create a text file to store the output audio files
+        os.chdir(reference_dir)
+        os.chdir("../")
 
-    with open('augmented_data/volume_reduction.txt', 'w') as file:
-        for item in output_files:
-            file.write(f"{item}\n")
+        with open('augmented_data/volume_reduction.txt', 'w') as file:
+            for item in output_files:
+                file.write(f"{item}\n")
 
-    # Cleanup: Merge the directories into one
-    current_path = os.getcwd() + "/augmented_data/"
-    make_directory(current_path + "volume_reduction/")
-    for path in directories_made:
-        for file in os.listdir(path):
-            shutil.move(path + file, current_path + "volume_reduction/" + file)
-        os.rmdir(path)
+    if len(vol_dBs) > 0:
+        # Cleanup: Merge the directories into one
+        current_path = os.getcwd() + "/augmented_data/"
+        make_directory(current_path + "volume_reduction/")
+        for path in directories_made:
+            for file in os.listdir(path):
+                shutil.move(path + file, current_path + "volume_reduction/" + file)
+            os.rmdir(path)
